% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGH30_genes}
\alias{MGH30_genes}
\title{3D tSNE of gene expression profile across cells from patient MGH30}
\format{Each row is a gene.}
\usage{
MGH30_genes
}
\description{
From Patel et al. 2014.
}
\keyword{datasets}

