% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{Glioblastoma Patient Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480 rows and 9 columns.}
\usage{
patients
}
\description{
From Patel et al. 2014.
}
\keyword{datasets}

