% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3plot.R
\name{qplot3}
\alias{qplot3}
\title{3D scatter plot on Rstudio viewer}
\usage{
qplot3(mappings, data, verbose_level = 0)
}
\description{
3D scatter plot on Rstudio viewer
}
\examples{
qplot3(aes(x=x, y=y, z=z, color=pick), ret)
}

