% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\docType{import}
\name{reexports}
\alias{aes}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}}
}}
\keyword{internal}

