% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ret}
\alias{ret}
\title{Test Data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 844 rows and 6 columns.}
\usage{
ret
}
\description{
Test Data set
}
\keyword{datasets}

