% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp.R
\name{zp}
\alias{zp}
\title{ZP}
\usage{
zp(data_, mappings_)
}
\description{
ZP
}
\examples{
data(patients)
zp(patients, zpa(x=pc1, y=pc2, z=pc3, color=patient))
zp(patients, list(pca=zpa(x=pc1, y=pc2, z=pc3, color=patient), mds=zpa(x=mds1, y=mds2, z=mds3, color=patient)))
}

