% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3plot.R
\name{qplot3}
\alias{qplot3}
\title{3D scatter plot on Rstudio viewer}
\usage{
qplot3(mappings, data, verbose_level = 0, show_datum = F)
}
\description{
3D scatter plot on Rstudio viewer
}
\examples{
library(ggplot2)
data(ret)
qplot3(aes(x=x, y=y, z=z, color=group), data=ret, show_datum=T)
}

