% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp.R
\name{zp}
\alias{zp}
\title{ZP}
\usage{
zp(data_, mapping_)
}
\description{
ZP
}
\examples{
data(patients)
zp(patients, list(x='pc1', y='pc2', z='pc3', color='patient'))
zp(patients, list(x='mds1', y='mds2', z='mds3', color='patient'))
}

