% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp.R
\name{zp}
\alias{zp}
\title{ZP}
\usage{
zp(data_, mappings_, use_viewer_ = T)
}
\description{
ZP
}
\examples{
data(patients)
data(MGH30genes)

zp(patients, list(PCA=zpa(x=pc1, y=pc2, z=pc3, color=patient),
                  MDS=zpa(x=mds1, y=mds2, z=mds3, color=patient),
                  MDS_expr=zpa(x=mds1, y=mds2, z=mds3, color=highly_expressed),
                  PCA_expr=zpa(x=pc1, y=pc2, z=pc3, color=highly_expressed)))

zp(MGH30genes, list(Correlation_tSNE=zpa(x=tsne1, y=tsne2, z=tsne3, color=pathway)))
}

