% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp.R
\name{zp}
\alias{zp}
\title{ZP}
\usage{
zp(data_, use_viewer = F)
}
\description{
ZP
}
\examples{
library(magrittr)

data(patients)
data(MGH30genes)

zp(patients) \%>\%
  zp_coord(pc1, pc2, pc3) \%>\%
  zp_coord(mds1, mds2, mds3) \%>\%
  zp_color(patient) \%>\%
  zp_color(avg_log_exp)

zp(MGH30_genes) \%>\%
  zp_coord(tsne1, tsne2, tsne3) \%>\%
  zp_color(pathway)
}

